backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
supabase
python-dotenv
